automation-server/requirements.txt
# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# 数据库
sqlalchemy==2.0.44
alembic==1.12.1
aiosqlite==0.19.0
greenlet>=3.0.0

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 工具库
pydantic>=2.10
pydantic-settings>=2.7
python-dotenv==1.0.0
Jinja2>=3.1
bcrypt>=4.0

# APK 解析
androguard==3.3.5
